Transplant using https://github.com/staticwebdev/blazor-starter
I could not figure out how to get the Azure Functions in the Api
project to respond. The example from [1] works, and I tried to compare
the differences between it and this solution. But I still could not
get this solution to work.

Not having any better idea, I transplanted my code into [1] and checked
that it worked in there. Then I just replaced whatever was in this
solution with what I got to work in [1]. This has the effect of nuking
away the authentication code that was not working anyway. At least I
can get the Azure Functions part working this way.

[1] https://github.com/staticwebdev/blazor-starter